Separated cli input types and operations from cli output types and operations
