+++ zork++/src/lib/cli/output/commands.rs
@@ -1,7 +1,6 @@
 ///! Contains helpers and data structure to process in
 /// a nice and neat way the commands generated to be executed
 /// by Zork++
-use std::process::Command;
 
 use crate::project_model::compiler::CppCompiler;
 use color_eyre::{eyre::Context, Result};
@@ -17,7 +16,7 @@ pub fn execute_command(compiler: &CppCompiler, arguments: &Vec<Argument<'_>>) ->
     );
 
     let process = if compiler.eq(&CppCompiler::MSVC) {
-        Command::new( // TODO The initialization process + cache process MUST dynamically get this path and store it in cache
+        std::process::Command::new( // TODO The initialization process + cache process MUST dynamically get this path and store it in cache
             "C:\\Program Files\\Microsoft Visual Studio\\2022\\Community\\VC\\Auxiliary\\Build\\vcvars64.bat"
         ).arg("&&")
             .arg(compiler.get_driver())
@@ -26,7 +25,7 @@ pub fn execute_command(compiler: &CppCompiler, arguments: &Vec<Argument<'_>>) ->
             .wait()
             .with_context(|| format!("[{compiler}] - Command {:?} failed!", arguments.join(" ")))?
     } else {
-        Command::new(compiler.get_driver())
+        std::process::Command::new(compiler.get_driver())
             .args(arguments)
             .spawn()?
             .wait()
